refactor: 拆分 CommunityAvatar 组件并优化布局
拆分：
- AnnouncementConstants.ts: 常量与工具函数（公告类型颜色、标签、时间格式化）
- AnnouncementList.tsx: 公告列表组件（展示公告摘要和未读指示）
- AnnouncementDetailDialog.tsx: 公告详情弹窗组件（Markdown 渲染）
- CommunityAvatarContent.tsx: 登录和未配置内容组件（LoggedInContent、NotConfiguredContent）
- CommunityAvatar.tsx: 主组件简化（聚焦状态管理和事件协调）

布局优化：
- 公告列表项调整：标题左侧，Chip（类型标签）和时间右侧竖排
- 提升代码可维护性和复用性
